move helpers to filters.js and update readme
